add https and dnd
