+++ .babelrc
@@ -9,5 +9,6 @@
         "react-hmre"
       ]
     }
-  }
+  },
+  "plugins":["transform-decorators-legacy"]
 }
