--- .npmrc
@@ -1,1 +1,0 @@
-optional=false
