+++ app/styles/header.css
@@ -12,6 +12,7 @@
 
 .header .logo {
   width: 300px;
+  height: 60px;
   pointer-events: none;
 }
