+++ app/styles/main.css
@@ -105,7 +105,6 @@ a {
   border: none;
   overflow: auto;
   outline: none;
-
   -webkit-box-shadow: none;
   -moz-box-shadow: none;
   box-shadow: none;
