+++ package.json
@@ -18,12 +18,12 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "bcryptjs": "^2.4.3",
     "bluebird": "^3.4.7",
     "body-parser": "^1.15.2",
     "compression": "^1.6.2",
     "cookie-parser": "^1.4.3",
-    "flash": "^1.1.0",
     "lodash": "^4.17.4",
     "method-override": "^2.3.6",
     "moment": "^2.18.1",
@@ -32,9 +32,9 @@
     "mongoose-paginate": "^5.0.3",
     "morgan": "^1.7.0",
     "nodemailer": "^3.1.8",
-    "passport": "^0.3.2",
-    "passport-local": "^1.0.0",
     "react": "^15.3.2",
+    "react-dnd": "^2.3.0",
+    "react-dnd-html5-backend": "^2.3.0",
     "react-dom": "^15.3.2",
     "react-google-maps": "^6.0.1",
     "react-router": "^3.0.2",
